requirements.txt
python-telegram-bot==21.6
paho-mqtt==1.6.1
